Move React components around.
