+++ app/frontend/components/Enrollment/Common/AtCapacity.jsx
