+++ app/frontend/components/Enrollment/LinkRendering.jsx
