+++ app/frontend/components/Enrollment/NewStudentError.jsx
