+++ app/frontend/components/Enrollment/StudentText.jsx
