+++ app/frontend/entrypoints/Common/home.jsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import ReactDOM from 'react-dom/client';
-import HomePage from '~/components/HomePage';
+import HomePage from '~/components/HomePage/HomePage';
 
 const container = document.getElementById('home');
 const root = ReactDOM.createRoot(container);
