+++ app/frontend/entrypoints/Students/confirmation.jsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import ReactDOM from 'react-dom/client';
-import ConfirmationPage from '~/components/ConfirmationPage';
+import ConfirmationPage from '~/components/ConfirmatonPage/ConfirmationPage';
 
 const container = document.getElementById('confirmation');
 const root = ReactDOM.createRoot(container);
