+++ app/frontend/entrypoints/Students/enrollments_index.jsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import ReactDOM from 'react-dom/client';
-import EnrollmentIndex from '~/components/EnrollmentIndex';
+import EnrollmentIndex from '~/components/Enrollment/EnrollmentIndex';
 
 const container = document.getElementById('index');
 const root = ReactDOM.createRoot(container);
